Cleanup README.md
Main change is in the formatting of the Java code so that it fits in the window. 

While at it, took out reference to DTLS-Fuzzer, as it is not interesting for the reader (esp. so early in the README text) and fixed non-existing link to EDHOC-Fuzzer.  These two deletions perhaps we can incorporate again, but in a more appropriate way.