Do not return mixed mutable/immutable results
A function returning either an immutable (`Collections.singletonList`) or
mutable (`LinkedList`) data structure may become quite confusing to its
callers.  Since the offending functions are intended mainly (only?) for
reading, this PR opted to make the results immutable (with a call to
`unmoodifiableList`).  It's possible that this conversion has a cost.
An alternative would have been to use some other construct instead of
`singetonList` for the simple case or a direct way to construct an
immutable list (using some incremental/buffering-like construction).